# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Donadomains API",
    "description": "Domain search aggregator — compare registration prices across 6 registrars (GoDaddy, Namecheap, Dynadot, Hover, Name.com, Porkbun) in real-time. Also provides detailed domain intelligence via RDAP and DNS.\n\nUse cases:\n- Find available domains for a keyword or brand name\n- Compare prices across registrars and find the cheapest\n- Check if a specific domain is taken\n- Look up domain ownership, expiry, registrar, nameservers\n\nNo authentication required. All prices are live-scraped, not cached.",
    "version": "1.0.0",
    "contact": {
      "name": "Donadomains",
      "url": "https://donadomains.xyz/docs"
    }
  },
  "externalDocs": {
    "description": "Full API documentation with interactive examples",
    "url": "https://donadomains.xyz/docs"
  },
  "servers": [
    {
      "url": "https://donadomains.xyz",
      "description": "Production"
    }
  ],
  "paths": {
    "/api/search": {
      "get": {
        "operationId": "searchDomains",
        "summary": "Search domain availability and compare prices across registrars",
        "description": "Scrapes 6 registrar search pages in parallel and returns merged results with live pricing. Each available domain includes buy links from every registrar that has it, with the cheapest option flagged. Taken domains are also returned (with empty buyLinks). Supports SSE streaming for progressive results.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Search keyword (e.g. \"startup\") or full domain (e.g. \"startup.io\"). A bare keyword returns results across all TLDs. Including a TLD prioritizes that exact domain in results.",
            "schema": { "type": "string", "minLength": 1, "maxLength": 63 },
            "examples": {
              "keyword": { "value": "myproject", "summary": "Search by keyword" },
              "domain": { "value": "myproject.io", "summary": "Search with specific TLD" }
            }
          },
          {
            "name": "stream",
            "in": "query",
            "required": false,
            "description": "Set to \"true\" for Server-Sent Events streaming. Results arrive progressively as each registrar completes (2-15 seconds). Recommended for real-time UIs.",
            "schema": { "type": "string", "enum": ["true"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results with pricing from multiple registrars",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SearchResponse" },
                "example": {
                  "keyword": "myproject",
                  "tld": null,
                  "results": [
                    {
                      "domain": "myproject.com",
                      "available": true,
                      "tld": ".com",
                      "matchType": "exact",
                      "buyLinks": [
                        { "name": "GoDaddy", "url": "https://www.godaddy.com/domainsearch/find?checkAvail=1&domainToCheck=myproject.com", "price": "$9.99/yr", "priceNum": 9.99, "isCheapest": true, "source": "scraped" },
                        { "name": "Dynadot", "url": "https://www.dynadot.com/domain/search?domain=myproject.com", "price": "$12.99/yr", "priceNum": 12.99, "source": "scraped" },
                        { "name": "Hover", "url": "https://www.hover.com/domains/results?q=myproject.com", "price": "$18.99/yr", "priceNum": 18.99, "source": "scraped" }
                      ],
                      "registerUrl": "https://www.godaddy.com/domainsearch/find?checkAvail=1&domainToCheck=myproject.com"
                    },
                    {
                      "domain": "myproject.io",
                      "available": true,
                      "tld": ".io",
                      "matchType": "exact",
                      "buyLinks": [
                        { "name": "Porkbun", "url": "https://porkbun.com/checkout/search?q=myproject.io", "price": "$28.12/yr", "priceNum": 28.12, "isCheapest": true, "source": "scraped" }
                      ],
                      "registerUrl": "https://porkbun.com/checkout/search?q=myproject.io"
                    },
                    {
                      "domain": "myproject.org",
                      "available": false,
                      "tld": ".org",
                      "matchType": "exact",
                      "buyLinks": []
                    }
                  ],
                  "sources": [
                    { "name": "Dynadot", "status": "ok", "count": 45, "durationMs": 6821 },
                    { "name": "Name.com", "status": "ok", "count": 27, "durationMs": 8234 },
                    { "name": "GoDaddy", "status": "ok", "count": 33, "durationMs": 12450 },
                    { "name": "Namecheap", "status": "ok", "count": 6, "durationMs": 9102 },
                    { "name": "Hover", "status": "ok", "count": 490, "durationMs": 11873 },
                    { "name": "Porkbun", "status": "failed", "count": 0, "durationMs": 20000, "error": "timeout" }
                  ]
                }
              },
              "text/event-stream": {
                "schema": { "type": "string", "description": "Server-Sent Events stream. Each event is a JSON object prefixed with 'data: '. Event types: init (registrar list), batch (results so far), complete (stream ends)." }
              }
            }
          },
          "400": {
            "description": "Missing or invalid query parameter",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "examples": {
                  "missing": { "value": { "error": "Missing q parameter", "example": "/api/search?q=example" } },
                  "invalid": { "value": { "error": "Invalid keyword" } }
                }
              }
            }
          }
        }
      }
    },
    "/api/domain/{domain}": {
      "get": {
        "operationId": "getDomainInfo",
        "summary": "Get detailed domain intelligence (RDAP + DNS)",
        "description": "Returns comprehensive domain intelligence for a single domain. Queries RDAP (registrar, dates, status, contacts) and DNS-over-HTTPS (nameservers, resolution) in parallel. Falls back to who.is scraping when RDAP data is sparse. Response time is typically 1-2 seconds.",
        "parameters": [
          {
            "name": "domain",
            "in": "path",
            "required": true,
            "description": "Fully qualified domain name. Must be lowercase alphanumeric with dots/hyphens, ending in a valid TLD.",
            "schema": { "type": "string", "pattern": "^[a-z0-9][a-z0-9.-]*[a-z0-9]\\.[a-z]{2,}$" },
            "examples": {
              "common": { "value": "github.com", "summary": "Well-known domain" },
              "cctld": { "value": "example.co.uk", "summary": "Country-code TLD" }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Domain intelligence",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DomainIntel" },
                "example": {
                  "domain": "github.com",
                  "registered": true,
                  "status": ["client delete prohibited", "client transfer prohibited", "client update prohibited"],
                  "created": "2007-10-09T18:20:50Z",
                  "updated": "2024-09-07T09:28:09Z",
                  "expires": "2025-10-09T18:20:50Z",
                  "registrar": "MarkMonitor Inc.",
                  "registrarUrl": "http://www.markmonitor.com",
                  "nameservers": ["dns1.p08.nsone.net", "dns2.p08.nsone.net", "dns3.p08.nsone.net", "dns4.p08.nsone.net", "ns-1283.awsdns-32.org", "ns-1707.awsdns-21.co.uk", "ns-421.awsdns-52.com", "ns-520.awsdns-01.net"],
                  "dnssec": "unsigned",
                  "sources": ["rdap", "dns"],
                  "timing": { "rdap": 287, "dns": 52 }
                }
              }
            }
          },
          "400": {
            "description": "Invalid domain format",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Invalid domain format" }
              }
            }
          },
          "504": {
            "description": "Request timed out (10 second limit)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "Gateway Timeout" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "SearchResponse": {
        "type": "object",
        "description": "Top-level search response containing merged results from all registrars.",
        "properties": {
          "keyword": { "type": "string", "description": "The parsed base keyword (e.g. \"myproject\" from \"myproject.io\")" },
          "tld": { "type": ["string", "null"], "description": "Extracted TLD if the query included one (e.g. \".io\"), null otherwise" },
          "results": {
            "type": "array",
            "description": "Domain results sorted by: available first, exact matches first, popular TLDs first, most registrar coverage first.",
            "items": { "$ref": "#/components/schemas/DomainResult" }
          },
          "sources": {
            "type": "array",
            "description": "Status of each registrar scrape — check this to see which registrars succeeded or failed.",
            "items": { "$ref": "#/components/schemas/SourceStatus" }
          }
        },
        "required": ["keyword", "results", "sources"]
      },
      "DomainResult": {
        "type": "object",
        "description": "A single domain with availability status and pricing from registrars.",
        "properties": {
          "domain": { "type": "string", "description": "Fully qualified domain name (e.g. \"myproject.com\")" },
          "available": { "type": "boolean", "description": "true if the domain can be registered, false if taken or premium" },
          "tld": { "type": "string", "description": "TLD including dot (e.g. \".com\", \".io\", \".xyz\")" },
          "matchType": { "type": "string", "enum": ["exact", "similar"], "description": "\"exact\" if domain is keyword+TLD, \"similar\" for suggestions (e.g. keyword variations)" },
          "buyLinks": {
            "type": "array",
            "description": "Purchase links from registrars, sorted cheapest first. Empty array for taken domains.",
            "items": { "$ref": "#/components/schemas/BuyLink" }
          },
          "registerUrl": { "type": "string", "description": "Direct URL to the cheapest registrar for quick purchase. Only present when buyLinks is non-empty." }
        },
        "required": ["domain", "available", "tld", "buyLinks"]
      },
      "BuyLink": {
        "type": "object",
        "description": "A purchase link from a single registrar with live pricing.",
        "properties": {
          "name": { "type": "string", "description": "Registrar name (e.g. \"GoDaddy\", \"Namecheap\", \"Dynadot\", \"Hover\", \"Name.com\", \"Porkbun\")" },
          "url": { "type": "string", "format": "uri", "description": "Direct purchase URL — takes the user to the registrar's checkout page" },
          "price": { "type": "string", "description": "Human-readable price (e.g. \"$9.99/yr\")" },
          "priceNum": { "type": "number", "description": "Numeric price in USD for sorting/comparison" },
          "isCheapest": { "type": "boolean", "description": "true if this is the lowest price across all registrars for this domain" },
          "premium": { "type": "boolean", "description": "true if this is a premium or aftermarket listing (not standard registration)" },
          "source": { "type": "string", "enum": ["scraped"], "description": "How the price was obtained. Always \"scraped\" (live from registrar website)." }
        },
        "required": ["name", "url", "source"]
      },
      "SourceStatus": {
        "type": "object",
        "description": "Status of a single registrar's scrape during the search.",
        "properties": {
          "name": { "type": "string", "description": "Registrar name" },
          "status": { "type": "string", "enum": ["ok", "failed"], "description": "\"ok\" if the registrar returned results, \"failed\" if scrape timed out or errored" },
          "count": { "type": "integer", "description": "Number of domain hits returned by this registrar (0 if failed)" },
          "durationMs": { "type": "integer", "description": "Time in milliseconds the registrar scrape took" },
          "error": { "type": "string", "description": "Error message if the registrar scrape failed (only present when status is \"failed\")" }
        },
        "required": ["name", "status", "count", "durationMs"]
      },
      "DomainIntel": {
        "type": "object",
        "description": "Comprehensive domain intelligence gathered from RDAP, DNS, and optionally who.is.",
        "properties": {
          "domain": { "type": "string", "description": "The normalized domain that was queried" },
          "registered": { "type": "boolean", "description": "true if the domain is currently registered, false if available" },
          "status": { "type": "array", "items": { "type": "string" }, "description": "ICANN domain status codes (e.g. \"client transfer prohibited\", \"client delete prohibited\")" },
          "created": { "type": "string", "format": "date-time", "description": "Domain creation/registration date (ISO 8601)" },
          "updated": { "type": "string", "format": "date-time", "description": "Last modification date (ISO 8601)" },
          "expires": { "type": "string", "format": "date-time", "description": "Expiration date (ISO 8601) — useful for monitoring domains that may become available" },
          "registrar": { "type": "string", "description": "Name of the registrar holding the domain (e.g. \"MarkMonitor Inc.\", \"GoDaddy.com, LLC\")" },
          "registrarUrl": { "type": "string", "format": "uri", "description": "Registrar's website URL" },
          "registrant": { "type": "string", "description": "Name of the domain registrant — often \"REDACTED FOR PRIVACY\" when privacy protection is enabled" },
          "organization": { "type": "string", "description": "Registrant's organization name" },
          "contactEmail": { "type": "string", "format": "email", "description": "Registrant contact email — only present when not privacy-protected" },
          "contactPhone": { "type": "string", "description": "Registrant contact phone" },
          "contactAddress": { "type": "string", "description": "Registrant postal address — usually partially or fully redacted" },
          "nameservers": { "type": "array", "items": { "type": "string" }, "description": "Authoritative DNS nameservers for the domain" },
          "dnssec": { "type": "string", "description": "DNSSEC status (e.g. \"signedDelegation\", \"unsigned\")" },
          "sources": { "type": "array", "items": { "type": "string", "enum": ["rdap", "dns", "who.is"] }, "description": "Which data sources were used to gather this intelligence" },
          "timing": { "type": "object", "additionalProperties": { "type": "number" }, "description": "Milliseconds spent querying each source (e.g. { \"rdap\": 342, \"dns\": 87 })" }
        },
        "required": ["domain", "registered", "sources", "timing"]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Human-readable error message" },
          "example": { "type": "string", "description": "Example of correct usage (included on some errors)" }
        },
        "required": ["error"]
      }
    }
  }
}